Update README.md
This HTML file serves as the entry point for a React.js-based weather app. It includes metadata for search engine optimization, links to the app's favicon and manifest file, and integrates Google Analytics for tracking. The body contains a div element with the id "root," where the React app will be mounted. The use of "%PUBLIC_URL%" in asset paths ensures compatibility with client-side routing. The provided script tags enable asynchronous loading of Google Analytics, setting up data tracking for the website. The app's functionality relies on JavaScript, and the user is prompted to enable it through a noscript tag.